Decouple all test helpers
